Add 'context' field for type hints. Fixes #51

+++ games/AxiomVerge2/src/prices.rs
@@ -8,9 +8,9 @@ use std::fmt;
 pub enum Currency {
     #[default]
     Free,
-    Energy(i32),
-    Flasks(i32),
-    Refills(i32),
+    Energy(i16),
+    Flasks(i8),
+    Refills(i8),
 }
 
 impl analyzer::world::Id for Currency {}
